refactor: `CellPathCompletion` to separated file
